Address comments:
- [x] move integration tests that require defvnet connection to an `integration` folder
- [x] add script to run just integration tests
- [x] move docs to Solana package
- [x] don't run integration tests in the ci
- [x] ignore all .env files in packages
