+++ .gitignore
@@ -14,3 +14,5 @@ dist
 packages/ethereum/cache
 packages/ethereum/hardhat.log
 packages/ethereum/hardhat.pid
+
+packages/**/.env
